+++ integration-tests/src/test/kotlin/org/stellar/anchor/platform/test/Sep31Tests.kt
@@ -20,6 +20,8 @@ import org.stellar.anchor.api.sep.sep31.Sep31GetTransactionResponse
 import org.stellar.anchor.api.sep.sep31.Sep31PostTransactionRequest
 import org.stellar.anchor.api.sep.sep31.Sep31PostTransactionResponse
 import org.stellar.anchor.apiclient.PlatformApiClient
+import org.stellar.anchor.apiclient.TransactionsOrderBy
+import org.stellar.anchor.apiclient.TransactionsSeps
 import org.stellar.anchor.auth.AuthHelper
 import org.stellar.anchor.platform.*
 import org.stellar.anchor.util.GsonUtils
