+++ platform/src/main/java/org/stellar/anchor/platform/controller/platform/PlatformController.java
@@ -8,6 +8,8 @@ import org.springframework.web.bind.annotation.*;
 import org.stellar.anchor.api.exception.AnchorException;
 import org.stellar.anchor.api.platform.*;
 import org.stellar.anchor.api.sep.SepTransactionStatus;
+import org.stellar.anchor.apiclient.TransactionsOrderBy;
+import org.stellar.anchor.apiclient.TransactionsSeps;
 import org.stellar.anchor.platform.service.TransactionService;
 import org.stellar.anchor.util.TransactionsParams;
